added rounded-a style "Alta"
Rob Carpenter from Oak Grove College in England
(http://oakgrovecollege.org.uk) designed a rounded-a to use in
OpenDyslexic and was very gracious to allow it to be shared with
everyone.

This means that development of the style was accelerated, and everyone
benefits by having this option available to them today, and as others
implement OpenDyslexic, it should trickle down into other apps, etc.

I'm sure you'll all join me in thanking Mr. Carpenter and Oak Grove
College in their contribution. :)